Address comments #1: Move duplicated adreno/mali logic to the shared folder.
